requirements.txt
langchain==0.3.9
langchain_groq==0.2.1
langchain_community==0.3.8
pypdf==5.1.0
sentence-transformers==3.2.1
faiss-cpu==1.9.0.post1
python-dotenv
